Adds the ability to record data to multiple metric endpoints, not just Bosun
To do so I've introduced a new property on `BosunOptions` called `Endpoints`. This is an array of `MetricEndpoint` instances that maps a name and a handler for a specific endpoint. A handler is an implementation of `IMetricHandler` that handles
serialization of metrics and their metadata and a way to flush that serialized data to the endpoint. This allows us to abstract the serialization format and the transport for an endpoint.

There are 5 implementations of `IMetricHandler` defined:
 - BosunMetricHandler - handles posting to Bosun like the previous functionality
 - DataDogMetricHandler - handles posting to DataDog’s REST API
 - DataDogStatsdMetricHandler - handles posting to DataDog's agent using statsd UDP
 - LocalMetricHanlder - keeps a snapshot of the latest metrics in memory so that they can be queried pulled rather than pushed
 - SignalFxMetricHandler - handles posting to SignalFX’s REST API

Along the way this tears out a bunch of old code that was originally intended to minimize the overhead of creating large byte arrays. Instead of pooling `Payload` objects we now depend on System.Buffers and Pipelines.Sockets.Unofficial so we can use the new `IBufferWriter<byte>` abstraction. This abstraction allows us to efficiently write to a non-continguous sequence of byte arrays, access them to flush their contents to an underlying transport and then give it back to the underlying writer so it can be re-used.

@mgravell has written an implementation of `IBufferWriter<T>` that automatically breaks our payloads down into chunks based upon the max payload specified in the relevant handler so we effectively get the same behaviour as before.

We maintain as many `IBufferWriter<byte>` instances as a handler needs to write to its underlying transport. E.g. Bosun needs a buffer for regular metrics, one for slowly changing metrics and one for metadata, whereas DataDog's statsd handler needs just a single buffer for all metrics.

At each snapshot interval we obtain the relevant buffer and serialise the metrics or metadata into it directly. For most metric systems that involves JSON serialisation so we now take a dependency on System.Text.Json which provides out-of-the-box support for serialising directly into an `IBufferWriter<byte>`.

At each flush interval we obtain each buffer and attempt to flush it. For each segment of bytes in the buffer we flush the contents to the underlying transport. In the event of failure to flush data to an endpoint we retry 3 times with a delay of 10 seconds between those retries before being dropped. Previously a failed payload would be tried indefinitely - this is a change in behaviour.

Other changes:
 - Removal of the BosunUrl and GetBosunUrl properties - URLs are injected by instantiating the relevant handler. GetBosunUrl has not yet been re-implemented
 - Removal of the MaxPayloadCount and MaxPayloadOption properties - these are provided by individual handlers.
 - Removal of the AccessToken and GetAccessToken properties - these have not yet been re-implemented
 - Removed the net45 build of the project, netstandard20 only