Adds the ability to record data to multiple metric endpoints, not just Bosun (#32)
* Adds the ability to record data to multiple metric endpoints, not just Bosun

To do so I've introduced a new property on `BosunOptions` called `Endpoints`. This is an array of `MetricEndpoint` instances that maps a name and a handler for a specific endpoint. A handler is an implementation of `IMetricHandler` that handles
serialization of metrics and their metadata and a way to flush that serialized data to the endpoint. This allows us to abstract the serialization format and the transport for an endpoint.

There are 5 implementations of `IMetricHandler` defined:
 - BosunMetricHandler - handles posting to Bosun like the previous functionality
 - DataDogMetricHandler - handles posting to DataDog’s REST API
 - DataDogStatsdMetricHandler - handles posting to DataDog's agent using statsd UDP
 - LocalMetricHanlder - keeps a snapshot of the latest metrics in memory so that they can be queried pulled rather than pushed
 - SignalFxMetricHandler - handles posting to SignalFX’s REST API

Along the way this tears out a bunch of old code that was originally intended to minimize the overhead of creating large byte arrays. Instead of pooling `Payload` objects we now depend on System.Buffers and Pipelines.Sockets.Unofficial so we can use the new `IBufferWriter<byte>` abstraction. This abstraction allows us to efficiently write to a non-continguous sequence of byte arrays, access them to flush their contents to an underlying transport and then give it back to the underlying writer so it can be re-used.

@mgravell has written an implementation of `IBufferWriter<T>` that automatically breaks our payloads down into chunks based upon the max payload specified in the relevant handler so we effectively get the same behaviour as before.

We maintain as many `IBufferWriter<byte>` instances as a handler needs to write to its underlying transport. E.g. Bosun needs a buffer for regular metrics, one for slowly changing metrics and one for metadata, whereas DataDog's statsd handler needs just a single buffer for all metrics.

At each snapshot interval we obtain the relevant buffer and serialise the metrics or metadata into it directly. For most metric systems that involves JSON serialisation so we now take a dependency on System.Text.Json which provides out-of-the-box support for serialising directly into an `IBufferWriter<byte>`.

At each flush interval we obtain each buffer and attempt to flush it. For each segment of bytes in the buffer we flush the contents to the underlying transport. In the event of failure to flush data to an endpoint we retry 3 times with a delay of 10 seconds between those retries before being dropped. Previously a failed payload would be tried indefinitely - this is a change in behaviour.

Other changes:
 - Removal of the BosunUrl and GetBosunUrl properties - URLs are injected by instantiating the relevant handler. GetBosunUrl has not yet been re-implemented
 - Removal of the MaxPayloadCount and MaxPayloadOption properties - these are provided by individual handlers.
 - Removal of the AccessToken and GetAccessToken properties - these have not yet been re-implemented
 - Removed the net45 build of the project, netstandard20 only

* `nameof` instead of `.ToString()`

* Kill multiple instances of Lazy

* Add MemoryExtensions and move all uses of MemoryMarsdhal.TryGetArray to it

* Fixes up how we handle payloads that go over multiple continguous blocks in a ReadOnlySequence

Now we keep track of each serialize operation that would cause us to go over our maximum payload size and use that to flag at what points we need flush our output to the underlying socket.

Also:

 - UDP socket usages now use `SocketAwaitableEventArgs` and repeatedly switch out the underlying buffers
 - a couple of minor typos in the SignalFX provider
 - moving some cheese around how we retrieve arrays from a `ReadOnlyMemory<byte>`
 - update the Pipelines.Sockets.Unofficial lib to latest stable.

Finally, add some benchmarks so we can see how much memory we're gobbling. Nothing shocking here, it's substantially lower than it used to be.

* Tweaks to allow the URIs of various handlers to be swapped out

* Lower-case rate for Bosun compatibility

* Don't bother to serialize or send metrics/metadata if the endpoint is not configured yet

* Fix up value serialization in DataDog's statsd handler

* Make sure we cleanup Socket and SocketAwaitableEventArgs when we have a problem sending a datagram

* Add some more in-depth error handling for JSON APIs

* Make sure SignalFx data is sent!

* Tweak how data is sent for DataDog

* Update the way we handle counters to reflect what happens in the majority of other metric platforms

Notably, Bosun expects a monotonically increasing value which it uses to calculate rates. This change makes it so that Bosun gets what it wants but that we pass just the delta to the other platforms

* Use s_colon rather than s_pipe in length calculations in DataDog statsd handler

* Fix up a number of metrics that don't play well in a world of multiple endpoints

Some metrics were being overwritten when `Serialize` was called which meant multiple calls would yield inconsistent results. These metrics now use `PreSerialize` to snapshot the value and then use that value when serializing.

* Change where `PreSerialize` is called so we don't keep snapshotting incorrect values

* Minor tweaks to fix up metrics for different providers

 - SignalFx expects timestamp in milliseconds since epoch
 - DRY up the `Trim` functionality used across providers
 - `LocalMetricHandler` should accumulate counter values and only clear when asked

* Fix test project to use latest C#

* Fix NRE in EventGauge

* Fix argument exception

* Make sure local metric handler doesn't call event handlers many times

* Moar testing in scratch app

* When under load and a queue gets full for a particular endpoint we weren't handling the exception so subsequent endpoints never got any metrics

* Fix up tracking of metrics on Bosun and Local handlers to work correctly with multiple instances of a metric with the same name but different tags

* Whitespace

* Add tracking of the endpoint used when invoking `AfterSerialization`

* Report loadsa metrics, like prod - trying to repro an issue

* Remove async shutdown

* Fix build

* Don't allocate AfterSerializationInfo unless we gotta

* Tweak NuGet settings

* Rename everything to be StackExchange.Metrics and setup builds

 - use Nerdbank versioning, fix builds to work with new scheme
- lots of renames - things that were called `Bosun` are now `Metrics` or `Metric`

* Files missed from last commit

* Fix up version.json

* Add GitHub actions build