Fix broken metrics sending
When multiple handlers were added to the collector and a handler in the middle of the array failed we would skip sending metrics to subsequent handlers. This commit fixes that bug by moving an exception handler inside a loop around the handlers and adds tests to verify everything is good.

It also adds the retry interval to be an option passed to the collector so that we can drastically reduce the runtime of tests when failures are induced!